ui(settings) - move code editor font size pref to the left pane
